Add custom actions picker, for tables for now
