+++ enterprise/frontend/src/metabase-enterprise/api/table-data-edit.ts
@@ -10,7 +10,11 @@ import type {
   TableUpdateRowsRequest,
   TableUpdateRowsResponse,
 } from "metabase-enterprise/data_editing/tables/types";
-import type { TableAction, WritebackAction } from "metabase-types/api/actions";
+import {
+  type DataGridWritebackAction,
+  TableAction,
+  WritebackAction,
+} from "metabase-types/api/actions";
 
 import { EnterpriseApi } from "./api";
 
@@ -77,14 +81,13 @@ export const tableDataEditApi = EnterpriseApi.injectEndpoints({
         },
       }),
     }),
-    getActions: builder.query<Array<WritebackAction | TableAction>, void>({
+    getActions: builder.query<DataGridWritebackAction[], void>({
       query: () => ({
         method: "GET",
         url: `/api/ee/data-editing/tmp-action`,
       }),
-      transformResponse: (response: {
-        actions: Array<WritebackAction | TableAction>;
-      }) => response?.actions,
+      transformResponse: (response: { actions: DataGridWritebackAction[] }) =>
+        response?.actions,
     }),
     executeAction: builder.mutation<
       TableExecuteActionResponse,
