+++ enterprise/frontend/src/metabase-enterprise/table-actions/settings/ActionPicker/index.ts
@@ -1,0 +1,1 @@
+export { ActionPicker } from "./ActionPicker";
