+++ enterprise/frontend/src/metabase-enterprise/table-actions/settings/AddOrEditActionSettingsContent/AddOrEditActionSettingsContent.module.css
@@ -68,7 +68,7 @@
   display: flex;
   flex: 0 0 auto;
   justify-content: flex-end;
-  padding-top: 1rem;
+  padding: 1rem;
   border-top: 1px solid var(--mb-color-border);
 }
