+++ enterprise/frontend/src/metabase-enterprise/table-actions/settings/AddOrEditActionSettingsContent/index.ts
@@ -1,0 +1,1 @@
+export { AddOrEditActionSettingsContent } from "./AddOrEditActionSettingsContent";
