+++ enterprise/frontend/src/metabase-enterprise/table-actions/settings/AddOrEditActionSettingsContent/types.ts
@@ -1,0 +1,5 @@
+import type { RowActionFieldSettings } from "metabase-types/api";
+
+export type ActionParametersFormValues = {
+  parameters: RowActionFieldSettings[];
+};
