+++ frontend/src/metabase/common/components/EntityPicker/components/EntityPickerModal/index.ts
@@ -1,1 +1,2 @@
 export * from "./EntityPickerModal";
+export * from "./EntityPickerContent";
